Add Stopwatch Implementation
Change GameAttempt time to Int
Implement Stopwatch return in MS

+++ Holdem Trainer/Holdem Trainer/Models/GameAttempt.swift
@@ -13,7 +13,7 @@ struct GameAttempt: Storable {
     
     var name: String
     var correct: Bool
-    var time: Float
+    var time: Int
 }
 
 // GameAttempt must confrom to Codable, FetchableRecord, MutablePersistableRecord protocols for functionality with GRDB
